NET-5: moved network controller interfaces to separate package to prevent import cycles
