+++ network/controller/common/options.go
@@ -14,7 +14,7 @@
  *    limitations under the License.
  */
 
-package controller
+package common
 
 import (
 	"time"
